+++ src/client/src/api/fetchApiV2.ts
@@ -88,6 +88,7 @@ async function handleFetchError(response: Response) {
         responseContent.detail || responseContent.message,
         response.status,
         responseContent.messageKey,
+        responseContent,
       );
     }
   }